Recommended changes made.
